Fix for when condtional module has __useDefault
If a module has __useDefault set to true, and exports anything which doesn't have conditionExport's value as a prop, the readMemberExpression function will return undefined. 
For example, I have a module that exports "browser", and has __useDefault set to true. The import promise will return just a string "browser". Than, readMemberExpression, looks for a prop call "default" on the "browser" string, once it is not found, it returns undefined. Later, causing a red herring error, as if the export is not resolving to a string. 
(On the way fixed the error message too).