Added AMD-style loader first syntax.
This is based on how AMD deals with loaders: https://github.com/amdjs/amdjs-api/blob/master/LoaderPlugins.md

Introduces the meta configuration option 'loaderFirst' which can be used to specify this style of loading. Useful for porting existing require.js code bases to System.js. This option tells System.js how to load files *within* a file specified in the meta config. This means it is most useful within the context of a package config which applies to an entire package.

Documentation will need to be updated. One test has been added but more extensive tests are probably needed for this to be a supported feature. Note that while the require.js version supports multiple loaders this has not been tested for the loaderFirst implementation in System.js.