I'm starting the process of cleaning this branch up so that my changes look more like the existing code. This means that in places where there are environmental differences, abstractions need to be created and incorporated into the existing code. I'll move away from having 'node-load' export a separate class.
Since some of the wrapper functions will be used by 'extras' and 'features', the 'extras' files have been converted to 'esm' format and Rollup is now used to bundle them into 'iife' format.

'features/import-map.js' has been refactored to work in both Node and the browser, but it still needs to be incorporated into the Node-centric code. That is next.